Refactor init.pp into several classes
Before this, all resources in the NTP module were in a single class.
In this patch, resources are split into ntp::install, ntp::configure,
ntp::service and ntp::params.

Note that we make two changes here to simulate the kinds of changes and
errors that might be introduced in a refactoring:

- Service name is incorrectly changed to ntpd
- Variable name is incorrectly typed as $server in params.pp